requirements.txt
asgiref==3.11.1
charset-normalizer==3.4.4
diff-match-patch==20241021
Django==5.1.15
django-extensions==4.1
django-filter==25.1
django-htmx==1.27.0
django-import-export==4.4.0
pillow==12.1.1
reportlab==4.4.10
sqlparse==0.5.5
tablib==3.9.0
tzdata==2025.3
